requirements.txt
paho-mqtt
blinkt
numpy
